Worker processes are split into many more separate tasks
